main: split out LobstersApp
